+++ Example/Emission/AppDelegate.m
@@ -28,6 +28,7 @@
 #import <Emission/ARConversationComponentViewController.h>
 #import <Emission/ARFairMoreInfoComponentViewController.h>
 #import <Emission/ARCityFairListComponentViewController.h>
+#import <Emission/ARCitySectionListComponentViewController.h>
 #import <Emission/ARFairComponentViewController.h>
 #import <Emission/ARFairBoothComponentViewController.h>
 #import <Emission/ARFairArtworksComponentViewController.h>
@@ -275,7 +276,7 @@ randomBOOL(void)
 
   BOOL isShow = [route hasPrefix:@"/show/"] || [route hasPrefix:@"show/"];
   BOOL isCityFairList = [route hasPrefix:@"/city-fair/"];
-
+  BOOL isCityList = [route hasPrefix:@"/city/"];
 
   if ([route hasPrefix:@"/artist/"] && [route componentsSeparatedByString:@"/"].count == 3) {
     NSString *artistID = [[route componentsSeparatedByString:@"/"] lastObject];
@@ -348,9 +349,15 @@ randomBOOL(void)
   } else if (isShow && [route hasSuffix:@"/info"]) {
     NSString *showID = [[route componentsSeparatedByString:@"/"] objectAtIndex:2];
     viewController = [[ARShowMoreInfoComponentViewController alloc] initWithShowID:showID];
+
   } else if (isCityFairList) {
     NSString *citySlug = [[route componentsSeparatedByString:@"/"] objectAtIndex:2];
     viewController = [[ARCityFairListComponentViewController alloc] initWithCitySlug:citySlug];
+
+  } else if (isCityList) {
+    NSString *citySlug = [[route componentsSeparatedByString:@"/"] objectAtIndex:2];
+    NSString *section = [[route componentsSeparatedByString:@"/"] objectAtIndex:3];
+    viewController = [[ARCitySectionListComponentViewController alloc] initWithCitySlug:citySlug section:section];
 
   } else if (isShow) {
     NSString *showID = [route componentsSeparatedByString:@"/"].lastObject;