+++ src/__generated__/MapRendererQuery.graphql.ts
@@ -31,6 +31,7 @@ query MapRendererQuery(
 fragment GlobalMap_viewer_3La17j on Viewer {
   city(slug: $citySlug) {
     name
+    slug
     sponsoredContent {
       introText
       artGuideUrl
@@ -249,7 +250,7 @@ return {
   "kind": "Request",
   "operationKind": "query",
   "name": "MapRendererQuery",
-  "id": "58706a5aed1c0e765b9ebbfb053ae7ba",
+  "id": "e770a63fb280561c7af1622c55d56d1b",
   "text": null,
   "metadata": {},
   "fragment": {
@@ -321,6 +322,13 @@ return {
             "plural": false,
             "selections": [
               v1,
+              {
+                "kind": "ScalarField",
+                "alias": null,
+                "name": "slug",
+                "args": null,
+                "storageKey": null
+              },
               {
                 "kind": "LinkedField",
                 "alias": null,