+++ src/__generated__/QueryRenderersCitySectionListQuery.graphql.ts
@@ -1,0 +1,435 @@
+/* tslint:disable */
+
+import { ConcreteRequest } from "relay-runtime";
+import { CitySectionList_city$ref } from "./CitySectionList_city.graphql";
+export type PartnerShowPartnerType = "GALLERY" | "MUSEUM" | "%future added value";
+export type QueryRenderersCitySectionListQueryVariables = {
+    readonly citySlug: string;
+    readonly partnerType?: PartnerShowPartnerType | null;
+};
+export type QueryRenderersCitySectionListQueryResponse = {
+    readonly city: ({
+        readonly " $fragmentRefs": CitySectionList_city$ref;
+    }) | null;
+};
+export type QueryRenderersCitySectionListQuery = {
+    readonly response: QueryRenderersCitySectionListQueryResponse;
+    readonly variables: QueryRenderersCitySectionListQueryVariables;
+};
+
+
+
+/*
+query QueryRenderersCitySectionListQuery(
+  $citySlug: String!
+  $partnerType: PartnerShowPartnerType
+) {
+  city(slug: $citySlug) {
+    ...CitySectionList_city_3XGJno
+  }
+}
+
+fragment CitySectionList_city_3XGJno on City {
+  name
+  shows(discoverable: true, first: 20, sort: START_AT_ASC, after: "", partnerType: $partnerType) {
+    pageInfo {
+      endCursor
+      hasNextPage
+    }
+    edges {
+      node {
+        id
+        _id
+        __id
+        is_followed
+        start_at
+        end_at
+        status
+        href
+        type
+        name
+        cover_image {
+          url
+        }
+        exhibition_period
+        partner {
+          __typename
+          ... on Partner {
+            name
+            type
+          }
+          ... on ExternalPartner {
+            name
+            __id
+          }
+          ... on Node {
+            __id
+          }
+        }
+        __typename
+      }
+      cursor
+    }
+  }
+}
+*/
+
+const node: ConcreteRequest = (function(){
+var v0 = [
+  {
+    "kind": "LocalArgument",
+    "name": "citySlug",
+    "type": "String!",
+    "defaultValue": null
+  },
+  {
+    "kind": "LocalArgument",
+    "name": "partnerType",
+    "type": "PartnerShowPartnerType",
+    "defaultValue": null
+  }
+],
+v1 = [
+  {
+    "kind": "Variable",
+    "name": "slug",
+    "variableName": "citySlug",
+    "type": "String"
+  }
+],
+v2 = {
+  "kind": "ScalarField",
+  "alias": null,
+  "name": "name",
+  "args": null,
+  "storageKey": null
+},
+v3 = {
+  "kind": "ScalarField",
+  "alias": null,
+  "name": "__id",
+  "args": null,
+  "storageKey": null
+},
+v4 = {
+  "kind": "ScalarField",
+  "alias": null,
+  "name": "type",
+  "args": null,
+  "storageKey": null
+},
+v5 = {
+  "kind": "ScalarField",
+  "alias": null,
+  "name": "__typename",
+  "args": null,
+  "storageKey": null
+};
+return {
+  "kind": "Request",
+  "operationKind": "query",
+  "name": "QueryRenderersCitySectionListQuery",
+  "id": "9d4b430528cc36488dd204733bfdc028",
+  "text": null,
+  "metadata": {},
+  "fragment": {
+    "kind": "Fragment",
+    "name": "QueryRenderersCitySectionListQuery",
+    "type": "Query",
+    "metadata": null,
+    "argumentDefinitions": v0,
+    "selections": [
+      {
+        "kind": "LinkedField",
+        "alias": null,
+        "name": "city",
+        "storageKey": null,
+        "args": v1,
+        "concreteType": "City",
+        "plural": false,
+        "selections": [
+          {
+            "kind": "FragmentSpread",
+            "name": "CitySectionList_city",
+            "args": [
+              {
+                "kind": "Variable",
+                "name": "partnerType",
+                "variableName": "partnerType",
+                "type": null
+              }
+            ]
+          }
+        ]
+      }
+    ]
+  },
+  "operation": {
+    "kind": "Operation",
+    "name": "QueryRenderersCitySectionListQuery",
+    "argumentDefinitions": v0,
+    "selections": [
+      {
+        "kind": "LinkedField",
+        "alias": null,
+        "name": "city",
+        "storageKey": null,
+        "args": v1,
+        "concreteType": "City",
+        "plural": false,
+        "selections": [
+          v2,
+          {
+            "kind": "LinkedField",
+            "alias": null,
+            "name": "shows",
+            "storageKey": null,
+            "args": [
+              {
+                "kind": "Literal",
+                "name": "after",
+                "value": "",
+                "type": "String"
+              },
+              {
+                "kind": "Literal",
+                "name": "discoverable",
+                "value": true,
+                "type": "Boolean"
+              },
+              {
+                "kind": "Literal",
+                "name": "first",
+                "value": 20,
+                "type": "Int"
+              },
+              {
+                "kind": "Variable",
+                "name": "partnerType",
+                "variableName": "partnerType",
+                "type": "PartnerShowPartnerType"
+              },
+              {
+                "kind": "Literal",
+                "name": "sort",
+                "value": "START_AT_ASC",
+                "type": "PartnerShowSorts"
+              }
+            ],
+            "concreteType": "ShowConnection",
+            "plural": false,
+            "selections": [
+              {
+                "kind": "LinkedField",
+                "alias": null,
+                "name": "pageInfo",
+                "storageKey": null,
+                "args": null,
+                "concreteType": "PageInfo",
+                "plural": false,
+                "selections": [
+                  {
+                    "kind": "ScalarField",
+                    "alias": null,
+                    "name": "endCursor",
+                    "args": null,
+                    "storageKey": null
+                  },
+                  {
+                    "kind": "ScalarField",
+                    "alias": null,
+                    "name": "hasNextPage",
+                    "args": null,
+                    "storageKey": null
+                  }
+                ]
+              },
+              {
+                "kind": "LinkedField",
+                "alias": null,
+                "name": "edges",
+                "storageKey": null,
+                "args": null,
+                "concreteType": "ShowEdge",
+                "plural": true,
+                "selections": [
+                  {
+                    "kind": "LinkedField",
+                    "alias": null,
+                    "name": "node",
+                    "storageKey": null,
+                    "args": null,
+                    "concreteType": "Show",
+                    "plural": false,
+                    "selections": [
+                      {
+                        "kind": "ScalarField",
+                        "alias": null,
+                        "name": "href",
+                        "args": null,
+                        "storageKey": null
+                      },
+                      {
+                        "kind": "ScalarField",
+                        "alias": null,
+                        "name": "id",
+                        "args": null,
+                        "storageKey": null
+                      },
+                      v3,
+                      {
+                        "kind": "ScalarField",
+                        "alias": null,
+                        "name": "is_followed",
+                        "args": null,
+                        "storageKey": null
+                      },
+                      {
+                        "kind": "ScalarField",
+                        "alias": null,
+                        "name": "start_at",
+                        "args": null,
+                        "storageKey": null
+                      },
+                      {
+                        "kind": "ScalarField",
+                        "alias": null,
+                        "name": "end_at",
+                        "args": null,
+                        "storageKey": null
+                      },
+                      {
+                        "kind": "ScalarField",
+                        "alias": null,
+                        "name": "status",
+                        "args": null,
+                        "storageKey": null
+                      },
+                      {
+                        "kind": "ScalarField",
+                        "alias": null,
+                        "name": "_id",
+                        "args": null,
+                        "storageKey": null
+                      },
+                      v4,
+                      v2,
+                      {
+                        "kind": "LinkedField",
+                        "alias": null,
+                        "name": "cover_image",
+                        "storageKey": null,
+                        "args": null,
+                        "concreteType": "Image",
+                        "plural": false,
+                        "selections": [
+                          {
+                            "kind": "ScalarField",
+                            "alias": null,
+                            "name": "url",
+                            "args": null,
+                            "storageKey": null
+                          }
+                        ]
+                      },
+                      {
+                        "kind": "ScalarField",
+                        "alias": null,
+                        "name": "exhibition_period",
+                        "args": null,
+                        "storageKey": null
+                      },
+                      {
+                        "kind": "LinkedField",
+                        "alias": null,
+                        "name": "partner",
+                        "storageKey": null,
+                        "args": null,
+                        "concreteType": null,
+                        "plural": false,
+                        "selections": [
+                          v5,
+                          v3,
+                          {
+                            "kind": "InlineFragment",
+                            "type": "ExternalPartner",
+                            "selections": [
+                              v2
+                            ]
+                          },
+                          {
+                            "kind": "InlineFragment",
+                            "type": "Partner",
+                            "selections": [
+                              v2,
+                              v4
+                            ]
+                          }
+                        ]
+                      },
+                      v5
+                    ]
+                  },
+                  {
+                    "kind": "ScalarField",
+                    "alias": null,
+                    "name": "cursor",
+                    "args": null,
+                    "storageKey": null
+                  }
+                ]
+              }
+            ]
+          },
+          {
+            "kind": "LinkedHandle",
+            "alias": null,
+            "name": "shows",
+            "args": [
+              {
+                "kind": "Literal",
+                "name": "after",
+                "value": "",
+                "type": "String"
+              },
+              {
+                "kind": "Literal",
+                "name": "discoverable",
+                "value": true,
+                "type": "Boolean"
+              },
+              {
+                "kind": "Literal",
+                "name": "first",
+                "value": 20,
+                "type": "Int"
+              },
+              {
+                "kind": "Variable",
+                "name": "partnerType",
+                "variableName": "partnerType",
+                "type": "PartnerShowPartnerType"
+              },
+              {
+                "kind": "Literal",
+                "name": "sort",
+                "value": "START_AT_ASC",
+                "type": "PartnerShowSorts"
+              }
+            ],
+            "handle": "connection",
+            "key": "CitySectionList_shows",
+            "filters": [
+              "discoverable",
+              "sort",
+              "partnerType"
+            ]
+          }
+        ]
+      }
+    ]
+  }
+};
+})();
+(node as any).hash = 'd48c8d1c2f6d11c09dcadd53eeda23c2';
+export default node;