+++ src/lib/Containers/index.tsx
@@ -1,4 +1,5 @@
 import CityFairList from "lib/Scenes/City/CityFairList"
+import CitySectionList from "lib/Scenes/City/CitySectionList"
 import { ShowArtistsContainer } from "../Components/Show/ShowArtists"
 import { ShowArtworksContainer } from "../Components/Show/ShowArtworks"
 import Fair from "../Scenes/Fair"
@@ -19,6 +20,7 @@ import WorksForYou from "./WorksForYou"
 export default {
   Artist,
   BidFlow,
+  CitySectionList,
   Conversation,
   CityFairList,
   Gene,