+++ src/lib/Scenes/Map/GlobalMap.tsx
@@ -233,12 +233,13 @@ export class GlobalMap extends React.Component<Props, State> {
     // TODO: map region filtering can live here.
     const filter = this.filters[this.state.activeIndex]
     const {
-      city: { name: cityName, sponsoredContent },
+      city: { name: cityName, slug: citySlug, sponsoredContent },
     } = this.props.viewer
     EventEmitter.dispatch("map:change", {
       filter,
       buckets: this.state.bucketResults,
       cityName,
+      citySlug,
       sponsoredContent,
       relay: this.props.relay,
     })
@@ -480,6 +481,7 @@ export const GlobalMapContainer = createFragmentContainer(
     fragment GlobalMap_viewer on Viewer @argumentDefinitions(citySlug: { type: "String!" }, maxInt: { type: "Int!" }) {
       city(slug: $citySlug) {
         name
+        slug
         sponsoredContent {
           introText
           artGuideUrl